+++ tests/python_client/conftest.py
@@ -10,7 +10,7 @@ from utils.util_log import test_log as log
 from base.client_base import param_info
 from check.param_check import ip_check, number_check
 from config.log_config import log_config
-from utils.utils import get_milvus, gen_unique_str, gen_default_fields, gen_binary_default_fields
+from utils.util_pymilvus import get_milvus, gen_unique_str, gen_default_fields, gen_binary_default_fields
 
 timeout = 60
 dimension = 128
