+++ tests/python_client/testcases/stability/test_mysql.py
@@ -7,7 +7,7 @@ from multiprocessing import Pool, Process
 import pytest
 from pymilvus import IndexType
 
-from utils.utils import get_milvus, gen_vectors, default_dim
+from utils.util_pymilvus import get_milvus, gen_vectors, default_dim
 from common.common_type import CaseLabel
