+++ tests/python_client/testcases/stability/test_restart.py
@@ -6,7 +6,7 @@ import logging
 import json
 from multiprocessing import Pool, Process
 import pytest
-from utils.utils import get_milvus, restart_server, gen_entities, gen_unique_str, default_nb
+from utils.util_pymilvus import get_milvus, restart_server, gen_entities, gen_unique_str, default_nb
 from common.constants import default_fields, default_entities
 from common.common_type import CaseLabel
