+++ tests/python_client/testcases/test_collection.py
@@ -8,7 +8,7 @@ from utils.util_log import test_log as log
 from common import common_func as cf
 from common import common_type as ct
 from common.common_type import CaseLabel, CheckTasks
-from utils.utils import *
+from utils.util_pymilvus import *
 from common import constants as cons
 
 prefix = "collection"
