+++ tests/python_client/testcases/test_connection.py
@@ -3,7 +3,7 @@ import concurrent.futures
 from pymilvus import DefaultConfig, Milvus
 
 from base.client_base import TestcaseBase
-from utils.utils import get_milvus, gen_invalid_ips, gen_invalid_ints, gen_invalid_uris
+from utils.util_pymilvus import get_milvus, gen_invalid_ips, gen_invalid_ints, gen_invalid_uris
 import common.common_type as ct
 import common.common_func as cf
 from common.code_mapping import ConnectionErrorMessage as cem
