+++ tests/python_client/testcases/test_query.py
@@ -11,7 +11,7 @@ from common import common_func as cf
 from common import common_type as ct
 from common.common_type import CaseLabel, CheckTasks
 from utils.util_log import test_log as log
-import utils.utils as ut
+import utils.util_pymilvus as ut
 
 prefix = "query"
 exp_res = "exp_res"
