+++ tests/python_client/testcases/test_search_20.py
@@ -6,7 +6,7 @@ from utils.util_log import test_log as log
 from common import common_func as cf
 from common import common_type as ct
 from common.common_type import CaseLabel, CheckTasks
-from utils.utils import *
+from utils.util_pymilvus import *
 from common.constants import *
 
 prefix = "search_collection"
